+++ libnbuss/BaseClient.h
@@ -22,8 +22,8 @@ public:
 	void close();
 
 	template <class T>
-	int write(const std::vector<T> &data) {
-		int data_size = data.size() * sizeof(T);
+	ssize_t write(const std::vector<T> &data) {
+		ssize_t data_size = data.size() * sizeof(T);
 		const char * p =  reinterpret_cast<const char*>(data.data());
 
 		LIB_LOG(debug) << "BaseClient::Write<> data_size = " << data_size << " sizeof(T)=" << sizeof(T);
@@ -31,9 +31,9 @@ public:
 		return write(p, data_size);
 	}
 
-	int write(const char * data, ssize_t data_size);
+	ssize_t write(const char * data, ssize_t data_size);
 
-	int write(const std::string &data);
+	ssize_t write(const std::string &data);
 
 	/**
 	 * read from socket, up to buffer_size bytes.
